pyproject.toml
[tool.poetry]
name = "chatgpt"
version = "0.1.0"
description = ""
authors = ["CDFMLR <cdfmlr@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
revchatgpt = "^2.3.0"
aiohttp = "^3.8.4"
protobuf = "^4.22.0"
grpcio = "^1.51.3"
grpcio-reflection = "^1.51.3"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
secondary = false


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
